APP_3(API en Render-GitHub)/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
email-validator==2.3.0
dnspython==2.8.0
fastapi-x402